+++ scripts/world/tile_art.gd
@@ -21,7 +21,7 @@ const ENTITY_KEYS := [
 	&"player", &"key", &"potion", &"coin", &"shard", &"bomb",
 	&"skeleton", &"crawler", &"wisp", &"shambler", &"golem",
 	&"archer", &"wraith", &"warden", &"bolt", &"relic", &"emblem",
-	&"aegis",
+	&"aegis", &"shrine",
 ]
 
 static var _tile_cache: Dictionary = {}
@@ -106,6 +106,8 @@ static func _entity_pattern(p_key: StringName) -> Array:
 			return _EMBLEM
 		&"aegis":
 			return _AEGIS
+		&"shrine":
+			return _SHRINE
 		&"key":
 			return _KEY
 		&"shard":
@@ -150,6 +152,8 @@ static func _entity_palette(p_key: StringName) -> Dictionary:
 			return { &"@": Color("#4a7ab5"), &"o": Color("#e8b84c"), &"+": Color("#dff0ff") }
 		&"aegis":
 			return { &"@": Color("#208060"), &"o": Color("#e8b84c"), &"+": Color("#7fffd4"), &"*": Color("#ffffff") }
+		&"shrine":
+			return { &"@": Color("#7147a8"), &"o": Color("#c58cff"), &"+": Color("#f0d9ff"), &"*": Color("#ffffff") }
 		&"key":
 			return { &"o": Color("#e8b84c"), &"@": Color("#b8860b"), &"+": Color("#fff0b0") }
 		&"shard":
@@ -382,6 +386,17 @@ const _AEGIS := [
 	"o@@++@@o",
 	"o@+**+@o",
 	"o@+**+@o",
+	".o@++@o.",
+	"..o@@o..",
+	"...oo...",
+	"........",
+]
+
+const _SHRINE := [
+	"...oo...",
+	"..o@@o..",
+	".o@++@o.",
+	".o@**@o.",
 	".o@++@o.",
 	"..o@@o..",
 	"...oo...",